requirements.txt
asgiref==3.11.0
Django==4.2.27
django-cors-headers==4.9.0
djangorestframework==3.16.1
gunicorn==23.0.0
packaging==26.0
psycopg2-binary==2.9.11
python-dotenv==1.2.1
sqlparse==0.5.5
typing_extensions==4.15.0
whitenoise==6.11.0
